requirements.txt
# Telegram
python-telegram-bot==21.7

# Notion
notion-client>=2.2.1

# AI (OpenRouter)
httpx>=0.27.0

# Voice (local)
faster-whisper>=1.0.0

# Config
python-dotenv>=1.0.0
